Rasterization

+++ utils/mat4.cpp
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <assert.h>
 
-mat4::mat4(float diagonal) 
+mat4::mat4(double diagonal) 
 {
 	for (int i = 0; i < 16; i++) 
 	{
@@ -17,10 +17,10 @@ mat4::mat4(float diagonal)
 	}
 }
 
-mat4::mat4(float e0, float e1, float e2, float e3, 
-           float e4, float e5, float e6, float e7, 
-           float e8, float e9, float e10, float e11, 
-           float e12, float e13, float e14, float e15) 
+mat4::mat4(double e0, double e1, double e2, double e3, 
+           double e4, double e5, double e6, double e7, 
+           double e8, double e9, double e10, double e11, 
+           double e12, double e13, double e14, double e15) 
 {
 	mat[0] = e0;
 	mat[1] = e1;
@@ -40,12 +40,12 @@ mat4::mat4(float e0, float e1, float e2, float e3,
 	mat[15] = e15;
 }
 
-float mat4::operator [](int index) const 
+double mat4::operator [](int index) const 
 {
 	assert(index >= 0 && index < 16);
 	return mat[index];
 }
-float& mat4::operator [](int index) 
+double& mat4::operator [](int index) 
 {
 	assert(index >= 0 && index < 16);
 	return mat[index];
@@ -60,7 +60,7 @@ mat4 mat4::operator *(const mat4& m) const
 	{
 		for (size_t j = 0; j < 4; j++)
 		{
-			float sum = 0;
+			double sum = 0;
 			for (size_t k = 0; k < 4; k++)
 			{
 				// For column-major order (OpenGL)
