+++ utils/vec3.h
@@ -2,13 +2,13 @@
 
 class vec3 {
 public:
-	float x, y, z;
+	double x, y, z;
 
 public:
 	vec3();
-	vec3(float x, float y, float z);
+	vec3(double x, double y, double z);
 
-	vec3& set(float x, float y, float z);
+	vec3& set(double x, double y, double z);
 
 	vec3& zero(void);
 
@@ -17,30 +17,30 @@ public:
 
 	vec3			operator+ (const vec3& v) const;
 	vec3			operator- (const vec3& v) const;
-	vec3			operator* (float scalar) const;
-	vec3			operator/ (float scalar) const;
+	vec3			operator* (double scalar) const;
+	vec3			operator/ (double scalar) const;
 
 	vec3& operator= (const vec3& v);
 	vec3& operator+=(const vec3& v);
 	vec3& operator-=(const vec3& v);
-	vec3& operator*=(float scalar);
-	vec3& operator/=(float scalar);
+	vec3& operator*=(double scalar);
+	vec3& operator/=(double scalar);
 
-	float& operator[](int index);
-	const float& operator[](int index) const;
+	double& operator[](int index);
+	const double& operator[](int index) const;
 
-	float			dot(const vec3& v) const;
+	double			dot(const vec3& v) const;
 	vec3			cross(const vec3& v) const;
 
-	float			magnitude(void) const;
+	double			magnitude(void) const;
 	vec3& normalize(void);
 
-	float* ptr(void);
-	const float* ptr(void) const;
+	double* ptr(void);
+	const double* ptr(void) const;
 
 public:
     // Used when scalar * vec3
     // Need friend to access private fields of vec3
-	friend vec3		operator*(float scalar, const vec3& v);
+	friend vec3		operator*(double scalar, const vec3& v);
 };
